[CDMI] Adjusted WPE-OpenCDM related recipes
This commit brings in a fix for a missing configuration field
in OpenCDM wpeframework-plugin. This was caused by multiplication
of flags, which served the same purpose.

Some bcm buildroot flags related to the CDMi were moved in order to
focus all flags related to decryption in one place.
Currently the software override decides, which CDMI options are
visible for selection (including the keybox & device certificate).