[vss-sdk] Temporary do not exclude libglib2
During installation, the libtool is trying to use
the libglib2 library from the build directory.
As we are stubbing libglib2 package, meaning there is
not library there, libtool fails. For some reason
it is not using the library from the sysroot.
This is a workaround and proper solution needs to
be implemented