Bringup VIP5202; Align WebKit and GStreamer sinks (#481)
This PR brings up the VIP5202 platform and aligns WebKit and GStreamer with
adaptive streaming improvements. The EME HEVC profile is enabled at WebKit level.

The new RDK community GStreamer sinks require the following header, which is
currently not available in the SDK. As a workaround, we have copied the
following headers into the build. Once the correct SDK is delivered this issue
will be gone.

Missing headers:
- BSEAV/lib/utils/biobits.h
- BSEAV/lib/utils/biobits_api.h

Additionally, to cope with closed-source patches,
https://github.com/Metrological/buildroot-external-acn repository has been
created and must be downloaded and configured before starting the build. That
repository will contain all closed-source patches required by the ACN platforms.

* WPEFrameworkPlugins: Use RDKCentral/ThunderNanoServices
* WPEFramework: Use RDKCentral/Thunder
* WPEFramework: Align config with current integration
* Update WebKit to include various playback improvements
* Upgrade gst1-bcm to RDK community version
* acn_5202_bcm72604_wpe_amz_defconfig: tune to current integration
* Move closed source patch to GLOBAL_PATCH_DIR
* Change GLOBAL_PATCH_DIR path
* Remove redundant WebKit patch
* Enable EME HEVC; Reorder patches;