Migration to GStreamer 1.18.6: gst-plugins-good: Revert "audioparsers: add back segment clipping to parsers that have lost it"
The reverted patch causes aacparse to clip buffers out of the GstSegment, so
when doing a seek back in MSE and reappending audio init-segment + data from
an earlier time, a new GstSegment isn't emitted for some reason and all the
GstBuffers earlier than the old (technically "current") GstSegment are clipped
by aacparse and lost. This causes important buffered ranges to be absent and
the MVT DASH shaka "35. DASH_DYNAMIC Seek" fails when using GStreamer 1.18.6.