gst1-plugins-good: Don't apply Manual-revert-of-bfd0e022-qtdemux-rework-segment-eve.patch on wpe-2.38
This commit refactors the wpe version-specific patches. Before, the
"Manual revert..." patch was always applied and then reverted for wpe-2.28,
(and should also be reverted for wpe-2.38 but wasn't being reverted, and this
commit solves that). Now it's never applied by default (so no need for revert
on wpe-2.28 and wpe-2.38) and only applyed on demand on wpe-2.22.